+++ src/main/java/lotto/constants/Prize.java
@@ -1,24 +1,24 @@
 package lotto.constants;
 
 public enum Prize {
-    ZERO(0, 0),
-    ONE(1, 0),
-    TWO(2, 0),
-    THREE(3, 5000),
-    FOUR(4, 50000),
-    FIVE(5, 1500000),
-    BONUS(5, 30000000),
-    SIX(6, 2000000000);
+    ZERO(0, 0L),
+    ONE(1, 0L),
+    TWO(2, 0L),
+    THREE(3, 5000L),
+    FOUR(4, 50000L),
+    FIVE(5, 1500000L),
+    BONUS(5, 30000000L),
+    SIX(6, 2000000000L);
 
     private final Integer match;
-    private final Integer payout;
+    private final Long payout;
 
-    Prize(Integer match, Integer payout) {
+    Prize(Integer match, Long payout) {
         this.match = match;
         this.payout = payout;
     }
 
-    public Integer getPayout() {
+    public Long getPayout() {
         return payout;
     }
